Add #ifdef LIBSERIALPORT_STATIC to libserialport.h
For building on Windows, this allows the creation of a library that can be linked statically. The project to build the library needs to be changed to "Static Library" under "General->Configuration Type", and LIBSERIALPORT_STATIC will need to be defined in the preprocessor settings. In addition, code using the library needs to define LIBSERIALPORT_STATIC.